+++ CHANGES.md
@@ -18,6 +18,7 @@ Release Notes.
 
 - Fix the wrong retention setting of each measure/stream/trace.
 - Fix server got panic when create/update property with high dist usage.
+- Fix incorrect key range update in sidx part metadata.
 
 ### Document